requirements.txt
asgiref==3.8.1
Django==5.1.4
djangorestframework==3.15.2
Faker==33.1.0
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
